Issue #948: extract tau-session crate
