+++ crates/tau-session/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "tau-session"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+anyhow = "1"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+tau-ai = { path = "../tau-ai" }
+
+[dev-dependencies]
+tempfile = "3"
